Reorganize into bossphorus.engine submodule
