+++ bossphorus/common/__init__.py
@@ -19,6 +19,11 @@ from typing import Tuple
 import attr
 
 
+class CutoutNotFoundError(Exception):
+    """Raise when a cutout cannot be found on the associated engine."""
+
+
+
 @attr.s(frozen=True)
 class CutoutCoordinateFrame:
     """Represents a specific BOSS cutout."""
